6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("Size of a cha: %1u byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %1u byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %1u byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %1u byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %1u byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a cha: %1u byte(s)\n",sizeof(a));
+	printf("Size of an int: %1u byte(s)\n", sizeof(b));
+	printf("Size of a long int: %1u byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %1u byte(s)\n", sizeof(d));
+	printf("Size of a float: %1u byte(s)\n", sizeof(f));
 	return (0);
 }
